backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pandas>=2.1
numpy>=1.26
openpyxl>=3.1
sentence-transformers>=2.6
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
pydantic>=2.6
python-multipart>=0.0.9
pyarrow>=15.0
pytest>=8.0
